+++ spec/CloudCodeLogger.spec.js
@@ -14,7 +14,6 @@ describe("Cloud Code Logger", () => {
         });
 
         Parse.Cloud.run('loggerTest').then(() => {
-            Parse.Cloud._removeHook('Functions', 'logTest');
             return logController.getLogs({from: Date.now() - 500, size: 1000});
         }).then((res) => {
             expect(res.length).not.toBe(0);
@@ -42,7 +41,6 @@ describe("Cloud Code Logger", () => {
 
         let obj = new Parse.Object('MyObject');
         obj.save().then(() => {
-            Parse.Cloud._removeHook('Triggers', 'beforeSave', 'MyObject');
             return logController.getLogs({from: Date.now() - 500, size: 1000})
         }).then((res) => {
             expect(res.length).not.toBe(0);
